+++ spring-data-geequery/src/main/java/com/github/geequery/springdata/annotation/IgnoreIf.java
@@ -6,9 +6,8 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * 用来描述JPA参数为某个指定的值的时候，参数不设置（忽略）不作为查询条件或更新字段
- * 
- *
+ * 用来描述参数为某个指定的值的时候，参数不设置（忽略）不作为查询条件或更新字段
+ * 配合GeeQuery中的NativeQuery子句自动省略功能。
  */
 @Target({ java.lang.annotation.ElementType.PARAMETER })
 @Retention(RetentionPolicy.RUNTIME)
