+++ spring-data-geequery/src/main/java/com/github/geequery/springdata/annotation/Modifying.java
@@ -23,19 +23,10 @@ import java.lang.annotation.Target;
 
 /**
  * Indicates a method should be regarded as modifying query.
- * 
- * @author Oliver Gierke
- * @author Christoph Strobl
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ ElementType.METHOD, ElementType.ANNOTATION_TYPE })
 @Documented
 public @interface Modifying {
-
-	/**
-	 * Defines whether we should clear the underlying persistence context after executing the modifying query.
-	 * 
-	 * @return
-	 */
 	boolean clearAutomatically() default false;
 }
