--- spring-data-geequery/src/main/java/com/github/geequery/springdata/repository/cdi/package-info.java
@@ -1,5 +1,0 @@
-/**
- * CDI support for Spring Data JPA Repositories. 
- */
-package com.github.geequery.springdata.repository.cdi;
-
