+++ spring-data-geequery/src/main/java/com/github/geequery/springdata/repository/config/package-info.java
@@ -1,5 +1,5 @@
 /**
- * Classes for JPA namespace configuration.
+ * Classes for GeeQuery namespace configuration.
  */
 package com.github.geequery.springdata.repository.config;
