+++ spring-data-geequery/src/main/java/com/github/geequery/springdata/repository/config/spring-gq.xsd
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8" ?>
-<xsd:schema xmlns="http://www.springframework.org/schema/data/jpa"
+<xsd:schema xmlns="http://www.springframework.org/schema/data/gq"
 	xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:tool="http://www.springframework.org/schema/tool"
 	xmlns:context="http://www.springframework.org/schema/context"
 	xmlns:repository="http://www.springframework.org/schema/data/repository"
