+++ spring-data-geequery/src/main/java/com/github/geequery/springdata/repository/query/GqNativeQuery.java
@@ -28,7 +28,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.orm.jpa.EntityManagerProxy;
 
-import com.github.geequery.springdata.annotation.IgnoreIf;
 import com.github.geequery.springdata.repository.query.GqParameters.GqParameter;
 
 /**
@@ -106,7 +105,7 @@ final class GqNativeQuery extends AbstractGqQuery {
 				continue;
 			}
 			
-			if (p.getIgnoreIf() != null && isIgnore(p.getIgnoreIf(), obj)) {
+			if (p.getIgnoreIf() != null && QueryUtils.isIgnore(p.getIgnoreIf(), obj)) {
 				continue;
 			}
 			if (p.isNamedParameter()) {
@@ -125,35 +124,6 @@ final class GqNativeQuery extends AbstractGqQuery {
 							+ "', please make sure that using @Param(\"name\") to mapping parameter into query.", e);
 				}
 			}
-		}
-	}
-
-	private boolean isIgnore(IgnoreIf ignoreIf, Object obj) {
-		switch (ignoreIf.value()) {
-		case Empty:
-			return obj == null || String.valueOf(obj).length() == 0;
-		case Negative:
-			if (obj instanceof Number) {
-				return ((Number) obj).longValue() < 0;
-			} else {
-				throw new IllegalArgumentException("can not calcuate is 'NEGATIVE' on parameter which is not a number.");
-			}
-		case Null:
-			return obj == null;
-		case Zero:
-			if (obj instanceof Number) {
-				return ((Number) obj).longValue() == 0;
-			} else {
-				throw new IllegalArgumentException("can not calcuate is 'IS_ZERO' on parameter which is not a number.");
-			}
-		case ZeroOrNagative:
-			if (obj instanceof Number) {
-				return ((Number) obj).longValue() <= 0;
-			} else {
-				throw new IllegalArgumentException("can not calcuate is 'IS_ZERO_OR_NEGATIVE' on parameter which is not a number.");
-			}
-		default:
-			throw new IllegalArgumentException("Unknown ignoreIf type:" + ignoreIf.value());
 		}
 	}
