--- spring-data-geequery/src/main/java/com/github/geequery/springdata/support/package-info.java
@@ -1,5 +1,0 @@
-/**
- * Various helper classes useful when working with JPA.
- */
-package com.github.geequery.springdata.support;
-
